code/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1

SQLAlchemy==2.0.34
alembic==1.13.2

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

stripe==10.12.0
httpx==0.27.2
reportlab==4.2.5
